requirements.txt
streamlit
torch
Pillow
numpy
basicsr
realesrgan
